requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Automat==25.4.16
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.3
cssselect==1.3.0
defusedxml==0.7.1
deprecation==2.1.0
filelock==3.18.0
frozenlist==1.6.0
gotrue==2.12.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
itemadapter==0.11.0
itemloaders==1.3.2
jmespath==1.0.1
lxml==5.4.0
multidict==6.4.3
packaging==25.0
parsel==1.10.0
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
Protego==0.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyDispatcher==2.0.7
PyJWT==2.10.1
pyOpenSSL==25.0.0
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
queuelib==1.8.0
realtime==2.4.3
requests==2.32.3
requests-file==2.1.0
Scrapy==2.13.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.1
supafunc==0.9.4
tldextract==5.3.0
Twisted==24.11.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
w3lib==2.3.1
websockets==14.2
yarl==1.20.0
zope.interface==7.2
